refactor: 패키지 오류로 인한 exception 클래스 패키지 이동
