+++ src/test/java/com/newzet/api/newsletter/business/NewsletterServiceConcurrencyTest.java
@@ -20,6 +20,7 @@ import com.newzet.api.common.cache.CacheUtil;
 import com.newzet.api.config.JwtTestConfig;
 import com.newzet.api.config.PostgresTestContainerConfig;
 import com.newzet.api.config.RedisTestContainerConfig;
+import com.newzet.api.newsletter.business.repository.NewsletterRepository;
 import com.newzet.api.newsletter.business.service.NewsletterService;
 import com.newzet.api.newsletter.fixture.NewsletterFixture;
 import com.newzet.api.newsletter.repository.NewsletterEntity;
