+++ src/test/java/com/newzet/api/newsletter/repository/NewsletterRepositoryImplTest.java
@@ -19,8 +19,8 @@ import com.newzet.api.category.repository.CategoryEntity;
 import com.newzet.api.category.repository.CategoryJpaRepository;
 import com.newzet.api.config.PostgresTestContainerConfig;
 import com.newzet.api.newsletter.business.dto.NewsletterEntityDto;
+import com.newzet.api.newsletter.exception.NoNewsletterException;
 import com.newzet.api.newsletter.fixture.NewsletterFixture;
-import com.newzet.api.newsletter.repository.exception.NoNewsletterException;
 
 @DataJpaTest
 @Import(NewsletterRepositoryImpl.class)
@@ -34,7 +34,6 @@ class NewsletterRepositoryImplTest {
 	private CategoryJpaRepository categoryJpaRepository;
 	@Autowired
 	private NewsletterJpaRepository newsletterJpaRepository;
-
 
 	private static void verifyFindByDomainOrMailingList(NewsletterEntityDto n1,
 		NewsletterEntityDto n2) {
@@ -148,7 +147,6 @@ class NewsletterRepositoryImplTest {
 		assertEquals(category.getId(), newsletter.getCategory().getId());
 	}
 
-
 	@Test
 	public void getNewsletterById() {
 		//Given
@@ -179,7 +177,7 @@ class NewsletterRepositoryImplTest {
 		//Given
 		List<NewsletterEntity> newsletterEntityList = new ArrayList<>();
 		List<UUID> categoryIdList = new ArrayList<>();
-		for (int i=0;i<5;i++) {
+		for (int i = 0; i < 5; i++) {
 			CategoryEntity category = categoryJpaRepository.save(
 				CategoryEntity.create("testCategory" + i, "test", "test"));
 			NewsletterEntity newsletter = newsletterJpaRepository.save(
@@ -193,8 +191,9 @@ class NewsletterRepositoryImplTest {
 			categoryIdList);
 
 		//Then
-		for (int i=0;i<5;i++) {
-			Assertions.assertThat(newsLetterListByCategoryIdList.get(i).getId()).isEqualTo(newsletterEntityList.get(i).getId());
+		for (int i = 0; i < 5; i++) {
+			Assertions.assertThat(newsLetterListByCategoryIdList.get(i).getId())
+				.isEqualTo(newsletterEntityList.get(i).getId());
 		}
 
 	}
